# pkg/delivery/http/openapi-v1.0.0.yml
openapi: 3.0.0
info:
  title: Midgard Public API
  version: 0.5.2
  contact:
    email: devs@thorchain.org
  description: The Midgard Public API queries THORChain and any chains linked via the Bifröst and prepares information about the network to be readily available for public users. The API parses transaction event data from THORChain and stores them in a time-series database to make time-dependent queries easy. Midgard does not hold critical information. To interact with BEPSwap and Asgardex, users should query THORChain directly.
paths:
...
  "/v1/txs":
    get:
      operationId: GetTxDetails
      summary: Get details of a tx by address, asset or tx-id
      description: Return an array containing the event details
      parameters:
        - in: query
          name: address
          description: Address of sender or recipient of any in/out tx in event
          required: false
          schema:
            type: string
          example: tbnb1fj2lqj8dvr5pumfchc7ntlfqd2v6zdxqwjewf5
        - in: query
          name: txid
          description: ID of any in/out tx in event
          required: false
          schema:
            type: string
          example: 2F624637DE179665BA3322B864DB9F30001FD37B4E0D22A0B6ECE6A5B078DAB4
        - in: query
          name: asset
          description: Any asset used in event (CHAIN.SYMBOL)
          required: false
          schema:
            type: string
          example: BNB.TOMOB-1E1
        - in: query
          name: type
          description: One or more comma separated unique types of event
          required: false
          schema:
            type: string
          example: [swap, stake, unstake, add, refund, doubleSwap]
        - in: query
          name: offset
          description: pagination offset
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
        - in: query
          name: limit
          description: pagination limit
          required: true
          schema:
            type: integer
            format: int64
            minimum: 0
            maximum: 50
      responses:
        "200":
          $ref: '#/components/responses/TxsResponse'
...
  "/v1/assets":
    get:
      operationId: GetAssetInfo
      summary: Get Asset Information
      description: Detailed information about a specific asset. Returns enough information to display a unique asset in various user interfaces, including latest price.
      parameters:
        - in: query
          name: asset
          description: One or more comma separated unique asset (CHAIN.SYMBOL)
          required: true
          schema:
            type: string
          example: [BNB.TOMOB-1E1, BNB.TCAN-014]
      responses:
        "400":
          $ref: '#/components/responses/GeneralErrorResponse'
        "200":
          $ref: '#/components/responses/AssetsDetailedResponse'
  "/v1/pools":
    get:
      operationId: GetPools
      summary: Get Asset Pools
      description: Returns an array containing all the assets supported on BEPSwap pools
      responses:
        "400":
          $ref: '#/components/responses/GeneralErrorResponse'
        "200":
          $ref: '#/components/responses/PoolsResponse'
  "/v1/pools/detail":
    get:
      operationId: GetPoolsDetails
      summary: Get Pools Details
      description: Returns an object containing all the pool details for that asset. 
      parameters:
        - in: query
          name: view
          description: Specifies the returning view
          schema:
            type: string
            enum: 
              - balances
              - simple
              - full
            default: full
        - in: query
          name: asset
          description: One or more comma separated unique asset (CHAIN.SYMBOL)
          required: true
          schema:
            type: string
          example: [BNB.TOMOB-1E1,BNB.TCAN-014]
      responses:
        "200":
          $ref: '#/components/responses/PoolsDetailedResponse'
...
  "/v1/stakers/{address}":
    get:
      operationId: GetStakersAddressData
      summary: Get Staker Data
      description: Returns an array containing all the pools the staker is staking in.
      parameters:
        - in: path
          name: address
          description: Unique staker address
          required: true
          schema:
            type: string
          example: 'bnb1jxfh2g85q3v0tdq56fnevx6xcxtcnhtsmcu64m'
      responses:
        "200":
          $ref: '#/components/responses/StakersAddressDataResponse'
  "/v1/stakers/{address}/pools":
    get:
      operationId: GetStakersAddressAndAssetData
      summary: Get Staker Pool Data
      description: Returns an object containing staking data for the specified staker and pool.
      parameters:
        - in: path
          name: address
          description: Unique staker address
          required: true
          schema:
            type: string
          example: 'bnb1jxfh2g85q3v0tdq56fnevx6xcxtcnhtsmcu64m'
        - in: query
          name: asset
          description: One or more comma separated unique asset (CHAIN.SYMBOL)
          required: true
          schema:
            type: string
          example: [BNB.TOMOB-1E1,BNB.TCAN-014]
      responses:
        "200":
          $ref: '#/components/responses/StakersAssetDataResponse'
...
  "/v1/history/total_volume":
    get:
      operationId: GetTotalVolChanges
      summary: Get Total Volume Changes
      description: Returns total volume changes of all pools in specified interval
      deprecated: true
      parameters:
        - in: query
          name: interval
          description: Interval of calculations
          required: true
          schema:
            type: string
            enum: ["5min", "hour", "day", "week", "month", "year"]
        - in: query
          name: from
          description: Start time of the query as unix timestamp
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: to
          description: End time of the query as unix timestamp
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: '#/components/responses/TotalVolChangesResponse'
  "/v1/history/pools":
    get:
      operationId: GetPoolAggChanges
      summary: Get Pool Aggregated Changes
      description: Returns historical aggregated details of the specified pool.
      parameters:
        - in: query
          name: pool
          description: Pool asset name
          required: true
          schema:
            type: string
          example: BNB.TOMOB-1E1
        - in: query
          name: interval
          description: Interval of calculations
          required: true
          schema:
            type: string
            enum: ["5min", "hour", "day", "week", "month", "year"]
        - in: query
          name: from
          description: Start time of the query as unix timestamp
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: to
          description: End time of the query as unix timestamp
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: '#/components/responses/GetPoolAggChangesResponse'
  "/v1/history/stats":
    get:
      operationId: GetStatsChanges
      summary: Get Stats Changes
      description: Returns stats and total changes of all pools in specified interval
      parameters:
        - in: query
          name: interval
          description: Interval of calculations
          required: true
          schema:
            type: string
            enum: ["5min", "hour", "day", "week", "month", "year"]
        - in: query
          name: from
          description: Start time of the query as unix timestamp
          required: true
          schema:
            type: integer
            format: int64
        - in: query
          name: to
          description: End time of the query as unix timestamp
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: '#/components/responses/StatsChangesResponse'
  "/v1/pools/details/{pool}/earnings":
    get:
      operationId: GetEarningDetail
      summary: Get Pool Earning Detail
      description: Returns earning report and apy of the specified pool.
      parameters:
        - in: path
          name: pool
          description: Pool asset name
          required: true
          schema:
            type: string
          example: BNB.BNB
      responses:
        "200":
          $ref: '#/components/responses/GetPoolEarningDetailResponse'
components:
  responses:
    PoolsDetailedResponse:
      description: object containing pool data for that asset
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/PoolDetail'
...
    TxsResponse:
      description: Returns an array of transactions
      content:
        application/json:
          schema:
            type: object
            properties:
              count:
                type: integer
                format: int64
              txs:
                type: array
                items:
                  $ref: '#/components/schemas/TxDetails'        
    StakersAddressDataResponse:
      description: array of all the pools the staker is staking in
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/StakersAddressData'
...
    StakersAssetDataResponse:
      description: object containing staking data for a specific staker and asset
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/StakersAssetData'
    AssetsDetailedResponse:
      description: object containing detailed asset information
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/AssetDetail'
    PoolsResponse:
      description: array of assets
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/asset'
    GeneralErrorResponse:
      description: error message
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TotalVolChangesResponse:
      description: Get Return an array of total volume changes.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/TotalVolChanges'
    GetPoolAggChangesResponse:
      description: Get Return an array of pool changes.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/PoolAggChanges'
    GetPoolEarningDetailResponse:
      description: Get Return an array of pool changes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PoolEarningDetail'
    StatsChangesResponse:
      description: Get Return an array of stats changes.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/StatsChanges'
  schemas:
...
    AssetDetail:
      type: object
      properties:
        asset:
          $ref: '#/components/schemas/asset'
        dateCreated:
          type: integer
          format: int64
        priceRune:
          type: string
    Error:
      required:
        - "error"
      type: object
      properties:
        error:
          type: string
...
    asset:
      type: string
...
servers:
  - url: http://127.0.0.1:8080
  - url: https://127.0.0.1:8080